Don't include <HsFFI.h> in template-hsc.h; fixes #2897
This works around the problem of hsc2hs not knowing the path to get it
from. We don't think many, if any, users actually need it. Those that do
will have to explicitly include it themselves, and their build system
will be responsible for passing a suitable -I flag to hsc2hs (Cabal
already does so).