Add alignment to hsc2hs template
Previously, calculating alignment values was commonly done by defining a
custom macro via `#let`:

```
(y__); }, y__)
```

Not only is this tedious, but it also doesn't work in cross-compilation
mode (see #10272). It makes sense to define this in `hsc2hs` for symmetry
with `#size`, `#poke`, `#peek`, etc., as well as to make it easier to
cross-compile `.hsc` files.

Fixes #4340.

Reviewed By: thomie, erikd

Differential Revision: https://phabricator.haskell.org/D1436

GHC Trac Issues: #4340, #10272