Build utilities with the bootstrap compiler when cross compiling
This should fix Trac #14297.  When building a cross compiler, we have rather little use
of utilities that do not run on the host, where the compiler runs. As such we should
build the utilities with the bootstrap (stage 0) compiler rather than witht he
in-tree (stage 1) compiler when CrossCompiling.

This used to results in the utilities we ship in the binary distribution to be built for
the wrong host.  This diff tries to rectify the situation and allow the binary distribution
to contain the utilities for the host when CrossCompiling.

See also https://phabricator.haskell.org/D4048