Adds a crude at&t assembly parser to resolve constants
Our current approach is taken from autoconf, and requires a binary search to
find constants.  This is not only very time consuming but also breaks when the
compiler fails to recognize an expression as constant.  As such we ask the
compiler to produce assembly, crudely parse that assembly and try to extract the
constant directly from the generated assembly.