Update travis.yml and tests
update travis.yml to not try to build the testsuite on 7.0 and 7.2
which won't work, because the dependencies 'call-stack' depends
on base >= 4.5, which is a dependency of hspec-core.

Also updates the assembly files, to drop the `\n` test, which might
end up as `\12`, as `\12`, `\x0a` and `\n`, testing this properly
would require some form of normalization, which we don't have yet.