Fix #35. Process arguments in order.
`foldl (.) id fs initial` processes list in reverse order, as it makes
`f1 (f2 (f3 ... initial ..))`, not intended `... (f3 (f2 (f1 initial)))`
application.

`addFlag` was therefore accumulating `cFlags` in reverse order.
Change that to use `snoc`.
This is not linear, but hopefully we are not passing 10000 flags
to hsc2hs.